Move POSIX implementation to shrs_lang
